Updated the comment about CXXFLAGS in make.conf. It shouldn't be uncommented by default if CFLAGS isn't.
Reorganized ebuild.sh to ensure that aliases are properly expanded... It
might be a problem in portage-2.0.47-r7 presently. Releasing -r8 shortly.
Just wanna be sure.

Repoman determines where it's at now. Doesn't assume it's in PORTDIR. It
figures out what PORTDIR should be and then reloads portage to get the
right results in the scan.

Fixed the --help reference.

svn path=/trunk/; revision=832