Skip modules the CPAN thinks are included with perl (closes bug 14679).
Used the CPAN module to discover the real location of Makefile.PL to set
the ${S} variable in the ebuild, sometimes the location isn't the same as
${P}.

Don't assume the filename of the tarball will be ${P}.tar.gz, use the
real filename out of CPAN.

Some modules' filenames have underscores in unfortunate places.  Change
all of them to hyphens to avoid that mess.

svn path=/trunk/; revision=893