Fix portage-multilib compilation of asm files (for autotools projects) when the user has set CCASFLAGS in make.conf.
During ./configure, if CCASFLAGS is unest it will default to CFLAGS
(or in essence, ${CCASFLAGS:=${CFLAGS}}). The old portage-multilib
behavior was to not define CCASFLAGS. However, certain users may wish
to define CCASFLAGS. In such a scenario, -m32 would never enter into
CCASFLAGS because the user's CCASFLAGS would be used by ./configure
instead of CFLAGS.

This patch causes portage-multilib to set CCASFLAGS to CFLAGS if it is
unset and then append the ABI-specific CFLAGS (such as -m32). This
allows a user to have different CCASFLAGS and still have successful
multilib builds.