portage/module.py: Add 'sourcefile' attribute to the module_spec
The file name otherwise had to be the same as the module name.
This only showed up when the module's target name (filename) was not the same as the
modules initialization name.
Add the new attribute to the emaint modules.
Add the new attribute to the the sync modules.
Make the plugin module parent name available.
Fix a docstring typo.